Adds priceRange to filter
