+++ components/Browse/AvailabilityFilters.tsx
@@ -2,28 +2,28 @@ import { Box } from "components/Box"
 import { Checkbox } from "components/Checkbox"
 import { Spacer } from "components/Spacer"
 import { Sans } from "components/Typography"
+import { color } from "helpers"
 import React from "react"
 import styled from "styled-components"
 
-export const AvailabilityFilters = ({ setParams, params }) => {
-  const { availableOnly, forSaleOnly } = params
+export const AvailabilityFilters = ({ setParams, params, breakpoint }) => {
+  const isMobile = breakpoint === "mobile"
+  const { available, forSaleOnly } = params
 
-  return (
-    <Box>
-      <Sans size="3">Filter by</Sans>
-      <Spacer mb={[0, 2]} />
-      <FlexWrapper mb={2}>
+  const Content = () => (
+    <>
+      <FlexWrapper mb={isMobile ? 0 : 2}>
         <Checkbox
           onClick={() => {
-            setParams({ ...params, availableOnly: !availableOnly })
+            setParams({ ...params, available: !available })
           }}
-          isActive={availableOnly}
+          isActive={available}
         />
         <Spacer ml={1} />
         <Sans size="3">Available now</Sans>
       </FlexWrapper>
-      <Spacer mb={[0, 2]} />
-      <FlexWrapper mb={5}>
+      <Spacer mb={isMobile ? 0 : 2} pr={isMobile ? 2 : 0} />
+      <FlexWrapper mb={isMobile ? 0 : 5}>
         <Checkbox
           onClick={() => {
             setParams({ ...params, forSaleOnly: !forSaleOnly })
@@ -33,7 +33,28 @@ export const AvailabilityFilters = ({ setParams, params }) => {
         <Spacer ml={1} />
         <Sans size="3">For sale</Sans>
       </FlexWrapper>
-    </Box>
+    </>
+  )
+
+  return (
+    <>
+      {isMobile ? (
+        <>
+          <Spacer mb={46} />
+          <FixedWrapper pt={1} pb={2} px={2}>
+            <FlexWrapper>
+              <Content />
+            </FlexWrapper>
+          </FixedWrapper>
+        </>
+      ) : (
+        <Box style={{ maxWidth: "148px" }}>
+          <Sans size="3">Filter by</Sans>
+          <Spacer mb={[0, 2]} />
+          <Content />
+        </Box>
+      )}
+    </>
   )
 }
 
@@ -41,4 +62,13 @@ const FlexWrapper = styled(Box)`
   display: flex;
   align-items: center;
   flex-direction: row;
+`
+
+const FixedWrapper = styled(Box)`
+  position: fixed;
+  top: 58px;
+  left: 0;
+  width: 100%;
+  z-index: 100;
+  background-color: ${color("white100")};
 `
